LayoutTests/fast/canvas/webgl/object-deletion-behaviour-expected.txt

Tests deletion behavior for texture, renderbuffer, shader, and program

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


shader and program deletion
PASS vertex shader loaded
PASS fragment shader loaded
PASS gl.attachShader(program, vertexShader) was expected value: NO_ERROR.
PASS gl.attachShader(program, fragmentShader) was expected value: NO_ERROR.
PASS gl.linkProgram(program) was expected value: NO_ERROR.
PASS gl.getProgramParameter(program, gl.LINK_STATUS) is true
PASS gl.useProgram(program) was expected value: NO_ERROR.
PASS gl.deleteShader(vertexShader) was expected value: NO_ERROR.
PASS gl.isShader(vertexShader) is true
PASS gl.getShaderParameter(vertexShader, gl.DELETE_STATUS) is true
PASS gl.detachShader(program, vertexShader) was expected value: NO_ERROR.
PASS gl.isShader(vertexShader) is false
PASS gl.deleteShader(fragmentShader) was expected value: NO_ERROR.
PASS gl.isShader(fragmentShader) is true
PASS gl.getShaderParameter(fragmentShader, gl.DELETE_STATUS) is true
PASS gl.deleteProgram(program) was expected value: NO_ERROR.
PASS gl.isProgram(program) is true
PASS gl.getProgramParameter(program, gl.DELETE_STATUS) is true
PASS gl.useProgram(null) was expected value: NO_ERROR.
PASS gl.isProgram(program) is false
PASS gl.isShader(fragmentShader) is false

texture deletion
PASS gl.bindFramebuffer(gl.FRAMEBUFFER, fbo) was expected value: NO_ERROR.
PASS gl.bindTexture(gl.TEXTURE_2D, tex) was expected value: NO_ERROR.
PASS gl.getParameter(gl.TEXTURE_BINDING_2D) is tex
PASS gl.framebufferTexture2D(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.TEXTURE_2D, tex, 0) was expected value: NO_ERROR.
PASS gl.getFramebufferAttachmentParameter(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.FRAMEBUFFER_ATTACHMENT_OBJECT_NAME) is tex
PASS gl.deleteTexture(tex) was expected value: NO_ERROR.
PASS gl.getFramebufferAttachmentParameter(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.FRAMEBUFFER_ATTACHMENT_OBJECT_NAME) is null
PASS gl.isTexture(tex) is false
PASS gl.getParameter(gl.TEXTURE_BINDING_2D) is null
PASS gl.bindTexture(gl.TEXTURE_2D, tex) was expected value: NO_ERROR.
PASS gl.getParameter(gl.TEXTURE_BINDING_2D) is null
PASS gl.bindTexture(gl.TEXTURE_CUBE_MAP, texCubeMap) was expected value: NO_ERROR.
PASS gl.getParameter(gl.TEXTURE_BINDING_CUBE_MAP) is texCubeMap
PASS gl.deleteTexture(texCubeMap) was expected value: NO_ERROR.
PASS gl.isTexture(texCubeMap) is false
PASS gl.getParameter(gl.TEXTURE_BINDING_CUBE_MAP) is null
PASS gl.bindTexture(gl.TEXTURE_CUBE_MAP, texCubeMap) was expected value: NO_ERROR.
PASS gl.getParameter(gl.TEXTURE_BINDING_CUBE_MAP) is null

renderbuffer deletion
PASS gl.bindRenderbuffer(gl.RENDERBUFFER, rbo) was expected value: NO_ERROR.
PASS gl.getParameter(gl.RENDERBUFFER_BINDING) is rbo
PASS gl.framebufferRenderbuffer(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.RENDERBUFFER, rbo) was expected value: NO_ERROR.
PASS gl.getFramebufferAttachmentParameter(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.FRAMEBUFFER_ATTACHMENT_OBJECT_NAME) is rbo
PASS gl.deleteRenderbuffer(rbo) was expected value: NO_ERROR.
PASS gl.getFramebufferAttachmentParameter(gl.FRAMEBUFFER, gl.COLOR_ATTACHMENT0, gl.FRAMEBUFFER_ATTACHMENT_OBJECT_NAME) is null
PASS gl.isRenderbuffer(rbo) is false
PASS gl.getParameter(gl.RENDERBUFFER_BINDING) is null
PASS gl.bindRenderbuffer(gl.RENDERBUFFER, rbo) was expected value: NO_ERROR.
PASS gl.getParameter(gl.RENDERBUFFER_BINDING) is null

buffer deletion
PASS gl.bindBuffer(gl.ARRAY_BUFFER, buffer) was expected value: NO_ERROR.
PASS gl.getParameter(gl.ARRAY_BUFFER_BINDING) is buffer
PASS gl.deleteBuffer(buffer) was expected value: NO_ERROR.
PASS gl.isBuffer(buffer) is false
PASS gl.getParameter(gl.ARRAY_BUFFER_BINDING) is null
PASS gl.bindBuffer(gl.ARRAY_BUFFER, buffer) was expected value: NO_ERROR.
PASS gl.getParameter(gl.ARRAY_BUFFER_BINDING) is null
PASS gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, bufferElement) was expected value: NO_ERROR.
PASS gl.getParameter(gl.ELEMENT_ARRAY_BUFFER_BINDING) is bufferElement
PASS gl.deleteBuffer(bufferElement) was expected value: NO_ERROR.
PASS gl.isBuffer(bufferElement) is false
PASS gl.getParameter(gl.ELEMENT_ARRAY_BUFFER_BINDING) is null
PASS gl.bindBuffer(gl.ELEMENT_ARRAY_BUFFER, bufferElement) was expected value: NO_ERROR.
PASS gl.getParameter(gl.ELEMENT_ARRAY_BUFFER_BINDING) is null

framebuffer deletion
PASS gl.bindFramebuffer(gl.FRAMEBUFFER, fbo) was expected value: NO_ERROR.
PASS gl.getParameter(gl.FRAMEBUFFER_BINDING) is fbo
PASS gl.deleteFramebuffer(fbo) was expected value: NO_ERROR.
PASS gl.isFramebuffer(fbo) is false
PASS gl.getParameter(gl.FRAMEBUFFER_BINDING) is null
PASS gl.bindFramebuffer(gl.FRAMEBUFFER, fbo) was expected value: NO_ERROR.
PASS gl.getParameter(gl.FRAMEBUFFER_BINDING) is null
PASS successfullyParsed is true

TEST COMPLETE
